--- linera-views/src/unit_tests/common_tests.rs
@@ -1,15 +1,0 @@
-// Copyright (c) Zefchain Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-use std::ops::Bound::{Excluded, Unbounded};
-
-use super::get_upper_bound;
-
-#[test]
-fn test_upper_bound() {
-    assert_eq!(get_upper_bound(&[255]), Unbounded);
-    assert_eq!(get_upper_bound(&[255, 255, 255, 255]), Unbounded);
-    assert_eq!(get_upper_bound(&[0, 2]), Excluded(vec![0, 3]));
-    assert_eq!(get_upper_bound(&[0, 255]), Excluded(vec![1]));
-    assert_eq!(get_upper_bound(&[255, 0]), Excluded(vec![255, 1]));
-}
