+++ linera-views/src/views/collection_view.rs
