+++ linera-views/src/views/hashable_wrapper.rs
