+++ linera-views/src/views/log_view.rs
