+++ linera-views/src/views/map_view.rs
