+++ linera-views/src/views/reentrant_collection_view.rs
