+++ linera-views/src/views/register_view.rs
